Refactor admin/public code
This splits the code so we can create a docker image without admin code.
