+++ api/pyproject.toml
@@ -22,3 +22,7 @@ exclude = '''
 
 [tool.isort]
 line_length=160
+multi_line_output=3
+include_trailing_comma=true
+force_grid_wrap=0
+use_parentheses=true
