+++ api/requirements/public.in
@@ -1,0 +1,14 @@
+connexion[swagger-ui]
+dockerflow
+flask
+flask_cors
+flask_migrate
+flask_talisman
+mozilla-version
+sentry-sdk[flask]
+slugid
+sqlalchemy
+
+# deployment
+gunicorn
+psycopg2
