+++ api/src/shipit_api/common/__init__.py
@@ -2,7 +2,3 @@
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-import shipit_api
-
-app = shipit_api.create_app()
