+++ api/tox.ini
@@ -46,8 +46,8 @@ commands =
     isort --check -rc -df {toxinidir}
     pip-compile-multi verify
     flake8 {toxinidir}
-    openapi-spec-validator {toxinidir}/src/shipit_api/api.yml
-    openapi-spec-validator {toxinidir}/src/shipit_api/api_public.yml
+    openapi-spec-validator {toxinidir}/src/shipit_api/admin/api.yml
+    openapi-spec-validator {toxinidir}/src/shipit_api/public/api.yml
 # Doesn't work without a .git in the same directory
 #    check-manifest -v {toxinidir}
